requirements.txt
alembic==1.13.2
attrs==23.2.0
black==24.4.2
blinker==1.8.2
cattrs==23.2.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
numpy==2.0.0
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.11
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
